requirements.txt
absl-py==1.3.0
adabelief-pytorch==0.2.1
anyio==3.4.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
attrs==21.4.0
auto-tqdm==1.0.3
Babel==2.9.1
backcall==0.2.0
bayesian-optimization==1.4.2
beautifulsoup4==4.11.1
bleach==4.1.0
branca==0.6.0
cachetools==5.0.0
celluloid==0.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
cmake==3.22.5
colorama==0.4.6
contourpy==1.0.6
cvxpy==1.2.1
cycler==0.11.0
data==0.4
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
derivative==0.4.2
dill==0.3.5.1
ecos==2.0.10
entrypoints==0.3
et-xmlfile==1.1.0
Fiona==1.8.22
flatbuffers==2.0
folium==0.14.0
fonttools==4.28.5
funcsigs==1.0.2
future==0.18.2
geodatasets==2023.3.0
geopandas==0.12.2
google-auth==2.6.0
google-auth-oauthlib==0.4.6
googledrivedownloader==0.4
GPUtil==1.4.0
graphviz==0.20
grpcio==1.44.0
h5py==3.1.0
haversine==2.8.0
idna==3.3
importlib-metadata==4.11.2
importlib-resources==5.4.0
ipykernel==6.6.0
ipython==7.30.1
ipython-genutils==0.2.0
isodate==0.6.1
jax==0.2.26
jedi==0.18.1
Jinja2==3.0.3
joblib==1.4.2
json5==0.9.6
jsonschema==4.3.3
jupyter==1.0.0
jupyter-client==7.1.0
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyter-server==1.13.1
jupyterlab==3.2.5
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.2
kiwisolver==1.3.2
latex==0.7.0
libpysal==4.7.0
mapclassify==2.4.3
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.6.2
matplotlib-inline==0.1.3
mistune==0.8.4
momepy==0.5.4
mpmath==1.2.1
munch==2.5.0
nbclassic==0.3.4
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==2.8.8
notebook==6.4.6
numpy==1.21.2
nvidia-ml-py3==7.352.0
oauthlib==3.2.0
olefile==0.46
opencv-python==4.6.0.66
openpyxl==3.1.2
opt-einsum==3.3.0
osmnx==1.2.2
osqp==0.6.2.post5
packaging==21.3
pandas==1.5.2
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
pikepdf==5.1.2
Pillow==9.0.0
pip==21.2.4
platformdirs==2.6.2
plotly==5.17.0
pooch==1.7.0
prometheus-client==0.12.0
prompt-toolkit==3.0.24
protobuf==3.19.4
psutil==5.9.1
ptyprocess==0.7.0
PuLP==2.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyDeprecate==0.3.1
pygad==3.3.1
pygeos==0.14
Pygments==2.11.1
pyparsing==3.0.6
pyproj==3.4.1
pyrsistent==0.18.0
pyshp==2.3.1
pysindy==1.7
pysr==0.6.14
python-dateutil==2.8.2
pytorch-lightning==1.4.0
pytorch-tcn==1.1.0
pytz==2021.3
PyYAML==6.0
pyzmq==22.3.0
qdldl==0.1.5.post2
rdflib==6.1.1
requests==2.28.1
requests-futures==1.0.0
requests-oauthlib==1.3.1
rsa==4.8
Rtree==1.0.1
scikit-learn==1.0.2
scipy==1.8.0
scs==3.2.0
seaborn==0.11.2
segregation==2.1.0
Send2Trash==1.8.0
setproctitle==1.3.2
Shapely==1.8.5.post1
shutilwhich==1.1.0
six==1.16.0
sklearn==0.0
sniffio==1.2.0
soupsieve==2.3.2.post1
statsmodels==0.13.1
sympy==1.9
tabulate==0.8.9
tempdir==0.7.1
tenacity==8.2.3
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
torch==1.12.1
torch-cluster==1.5.9
torch-geometric==2.0.3
torch-scatter==2.0.9
torch-sparse==0.6.18
torch-spline-conv==1.2.1
torch-summary==1.4.5
torchaudio==0.12.1
torchfile==0.1.0
torchmetrics==0.9.2
torchvision==0.13.1
torchviz==0.0.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typing_extensions==4.0.1
urllib3==1.26.7
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
Werkzeug==2.0.3
yarl==1.7.2
zipp==3.7.0
